requirements.txt
pyyaml>=6.0
markdown>=3.4.1
jinja2>=3.0.0
pygments>=2.13.0
python-frontmatter>=1.0.0
pathlib>=1.0.1
argparse>=1.4.0
requests>=2.28.0